pyproject.toml
[project]
name = "mcp-hub-project"
version = "0.2.0"
description = "Advanced MCP Hub with Inter-Agent Communication and Performance Monitoring"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio-client>=1.10.2",
    "gradio[mcp]>=5.33.0",
    "modal>=1.0.2",
    "openai>=1.84.0",
    "tavily-python>=0.7.4",
    "python-dotenv>=1.0.0",
    "psutil>=5.9.0",
    "aiohttp>=3.8.0",
    "anthropic>=0.52.2",
    "huggingface>=0.0.1",
    "huggingface-hub>=0.32.4",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
]
